fix dynamic rendering of variables
